chore: simplify the format
