+++ utils/mochaRunner/src/test.ts
@@ -1,16 +1,11 @@
 import assert from 'assert/strict';
 import test from 'node:test';
 import {
-  getExpectationsForModifiers,
+  filterByParameters as filterByParameters,
   getTestResultForFailure,
-} from './expectations.js';
-import {TestExpectation} from './types.js';
-import {
-  getFilename,
-  isEqualObject,
-  extendProcessEnv,
-  traverseMatrix,
 } from './utils.js';
+import {TestExpectation} from './types.js';
+import {getFilename, extendProcessEnv} from './utils.js';
 
 test('extendProcessEnv', () => {
   const env = extendProcessEnv([{TEST: 'TEST'}, {TEST2: 'TEST2'}]);
@@ -23,26 +18,6 @@ test('getFilename', () => {
   assert.equal(getFilename('/etc/test.js'), 'test');
 });
 
-test('traverseMatrix', () => {
-  const combinations = [];
-  for (const combination of traverseMatrix(
-    [],
-    [
-      [{A: true}, {B: true}],
-      [{C: true}, {D: true}],
-    ]
-  )) {
-    combinations.push(combination);
-  }
-  assert.equal(combinations.length, 4);
-  assert.deepEqual(combinations, [
-    [{A: true}, {C: true}],
-    [{A: true}, {D: true}],
-    [{B: true}, {C: true}],
-    [{B: true}, {D: true}],
-  ]);
-});
-
 test('getTestResultForFailure', () => {
   assert.equal(
     getTestResultForFailure({err: {code: 'ERR_MOCHA_TIMEOUT'}}),
@@ -51,68 +26,24 @@ test('getTestResultForFailure', () => {
   assert.equal(getTestResultForFailure({err: {code: 'ERROR'}}), 'FAIL');
 });
 
-test('isEqualObject', () => {
-  assert.equal(isEqualObject({a: 'b', c: 'd'}, {c: 'd', a: 'b'}), true);
-  assert.equal(isEqualObject({a: 'b'}, {c: 'd', a: 'b'}), false);
-});
-
-test('getExpectationsForModifiers', () => {
+test('filterByParameters', () => {
   const expectations: TestExpectation[] = [
     {
-      fullTitle:
-        'OOPIF "after all" hook for "should keep track of a frames OOP state"',
-      filename: 'oopif.spec',
+      testIdPattern:
+        '[oopif.spec] OOPIF "after all" hook for "should keep track of a frames OOP state"',
       platforms: ['darwin'],
-      modifiers: [
-        {
-          PUPPETEER_PRODUCT: 'firefox',
-        },
-        {
-          HEADLESS: 'true',
-        },
-      ],
+      parameters: ['firefox', 'headless'],
       expectations: ['FAIL'],
     },
   ];
   assert.equal(
-    getExpectationsForModifiers(expectations, [
-      {
-        PUPPETEER_PRODUCT: 'firefox',
-      },
-      {
-        HEADLESS: 'true',
-      },
-    ]).length,
+    filterByParameters(expectations, ['firefox', 'headless']).length,
     1
   );
+  assert.equal(filterByParameters(expectations, ['firefox']).length, 0);
   assert.equal(
-    getExpectationsForModifiers(expectations, [
-      {
-        PUPPETEER_PRODUCT: 'firefox',
-      },
-    ]).length,
-    0
-  );
-  assert.equal(
-    getExpectationsForModifiers(expectations, [
-      {
-        PUPPETEER_PRODUCT: 'firefox',
-      },
-      {
-        HEADLESS: 'true',
-      },
-      {
-        SMTH: 'ELSE',
-      },
-    ]).length,
+    filterByParameters(expectations, ['firefox', 'headless', 'other']).length,
     1
   );
-  assert.equal(
-    getExpectationsForModifiers(expectations, [
-      {
-        SMTH: 'ELSE',
-      },
-    ]).length,
-    0
-  );
+  assert.equal(filterByParameters(expectations, ['other']).length, 0);
 });
